parser/pyproject.toml
[tool.poetry]
name = "steam_parser_server"
version = "0.1.0"
description = ""
authors = ["Aleksei Monakhov <lexmonakhov@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.3.3"
selenium = "^4.20"
python-dotenv = "^0.20.0"
requests = "^2.26"
psycopg2-binary = "^2.9"

[tool.poetry.dev-dependencies]
pytest = "^6.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
